feat: move MultiColumnLayout to Shared component and update Home to use new layout structure
